Fixed auto connect retrying when unplugged, added physical connect listener
- Auto connect was retrying all 5 attempts and opening the dialog when
there were no known devices (flipper unplugged). It now stops after that
check.
- Added an event listener to the navigator serial for 'connect'. Before
we only had 'disconnect', but now if the auto connect toggle is ON we
will trigger a connect, making it more seamless.
- Removed auto connect from the home page as theres nothing strictly
needing the flipper connection on the home page (like in the updater or
asset pack pages).
- Grouped lines in the logs component in the updater in case multiple
logs of the same message happen.