Remove langs (again), small wiki fixes
- Localization is cool, and we can add it back later since the configs
are still built dynamically, but it's just not worth the effort right
now. Especially if things are added/changed in the wiki and all langs
need to follow. It also doesn't really make sense for the site as a
whole to have translation but the wiki just stay in English... so yeah.
- Fixed some headers and replaced the `*` with `Default` for all the
default options in the Momentum Settings pages.
- Updates to the `README`'s to reflect the langs removal and changed
`ballerina` to `zig` for some of the codeblocks because the colors were
off.
- Added "I forgot my PIN, how do I reset it?" to `FAQ`